+++ visualization/app/codeCharta/state/store/appSettings/enableExperimentalFeatures/experimentalFeaturesEnabled.splitter.ts
@@ -1,5 +1,0 @@
-import { setExperimentalFeaturesEnabled } from "./experimentalFeaturesEnabled.actions"
-
-export function splitExperimentalFeaturesEnabledAction(payload: boolean) {
-	return setExperimentalFeaturesEnabled(payload)
-}