+++ visualization/app/codeCharta/state/store/dynamicSettings/colorMode/colorMode.splitter.ts
@@ -1,6 +1,0 @@
-import { ColorModeAction, setColorMode } from "./colorMode.actions"
-import { ColorMode } from "../../../../codeCharta.model"
-
-export function splitColorModeAction(payload: ColorMode): ColorModeAction {
-	return setColorMode(payload)
-}