+++ visualization/app/codeCharta/state/store/dynamicSettings/edgeMetric/edgeMetric.splitter.ts
@@ -1,5 +1,0 @@
-import { setEdgeMetric } from "./edgeMetric.actions"
-
-export function splitEdgeMetricAction(payload: string) {
-	return setEdgeMetric(payload)
-}