+++ visualization/app/codeCharta/state/store/dynamicSettings/searchPattern/searchPattern.splitter.ts
@@ -1,5 +1,0 @@
-import { setSearchPattern } from "./searchPattern.actions"
-
-export function splitSearchPatternAction(payload: string) {
-	return setSearchPattern(payload)
-}