+++ visualization/app/codeCharta/state/store/fileSettings/markedPackages/markedPackages.splitter.ts
@@ -1,6 +1,0 @@
-import { setMarkedPackages } from "./markedPackages.actions"
-import { MarkedPackage } from "../../../../codeCharta.model"
-
-export function splitMarkedPackagesAction(payload: MarkedPackage[]) {
-	return setMarkedPackages(payload)
-}